wordcloud/requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.0
nltk==3.3
numpy==1.15.2
pandas==0.23.4
Pillow==5.3.0
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
wordcloud==1.5.0
